add a safety net for errnos raised on successful removal and log errors
for unknown reasons sometimes it happens on the arm node that the first rm
raises, we chown and then the second rm fails because it says the dir
isn't there anymore. prevent this by only running the fallback iff the
dir still exists in the errno block.

for future reference also log the error before chowning as to allow more
focused diagnoses should something similar happen again

+++ ci-tooling/nci/workspace_cleaner.rb
@@ -59,7 +59,12 @@ module WorkspaceCleaner
     def rm_r(dir)
       FileUtils.rm_r(dir, verbose: true)
     rescue SystemCallError => e
+      if File.exist?(dir)
+        warn "  Got error #{e} but still successfully removed directory."
+        return
+      end
       raise e unless e.class.name.start_with?('Errno::')
+      warn "  Got error #{e} trying to chown."
       chown_r(dir)
       FileUtils.rm_r(dir, verbose: true)
     end
